dom/base/test/test_bug590812.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for XML pretty printing, bug 590812</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=590812">Mozilla Bug 590812</a>
<p id="display"></p>
<iframe id=iframe></iframe>
<iframe src="file_bug590812-ref.xhtml"></iframe>
<pre id="test">
<script class="testbody" type="application/javascript">

add_task(async function start() {
  var noxul = "https://sub1.test1.example.com:443";
  var yesxul = "https://example.org:443"

  await SpecialPowers.pushPermissions([
    { type: "allowXULXBL", allow: false, context: noxul },
    { type: "allowXULXBL", allow: true, context: yesxul }
  ]);

  var path = "/tests/dom/base/test/file_bug590812.xml";
  var iframe = $('iframe');
  iframe.src = noxul + path;
  await new Promise(resolve => iframe.addEventListener("load", resolve, {once: true}));
  let sNoXUL = await snapshotWindow(window.frames[0], false);

  iframe.src = yesxul + path;
  await new Promise(resolve => iframe.addEventListener("load", resolve, {once: true}));
  let sWithXUL = await snapshotWindow(window.frames[0], false);

  let sRef = await snapshotWindow(window.frames[1], false);

  let res;
  ok(compareSnapshots(sNoXUL, sRef, true)[0],
     "noxul domain same as ref");
  ok(compareSnapshots(sWithXUL, sRef, true)[0],
     "xul supporting domain same as ref");
});

</script>
</pre>
</body>
</html>
